\name{constellation}
\alias{constellation}
\docType{data}
\title{
Constellation Lines Data
}
\description{
Constellation Lines Data
}
\usage{data("constellation")}
\format{
  A data frame with 1126 observations on the following 5 variables.
  \describe{
    \item{\code{CON}}{a factor with levels \code{} \code{AND} \code{ANT} \code{APS} \code{AQL} \code{AQR} \code{ARA} \code{ARI} \code{AUR} \code{BOO} \code{CAE} \code{CAM} \code{CAP} \code{CAR} \code{CAS} \code{CEN} \code{CEP} \code{CET} \code{CHA} \code{CIR} \code{CMA} \code{CMI} \code{CNC} \code{COL} \code{COM} \code{CRA} \code{CRB} \code{CRT} \code{CRU} \code{CRV} \code{CVN} \code{CYG} \code{DEL} \code{DOR} \code{DRA} \code{EQU} \code{ERI} \code{FOR} \code{GEM} \code{GRU} \code{HER} \code{HOR} \code{HYA} \code{HYI} \code{IND} \code{LAC} \code{LEO} \code{LEP} \code{LIB} \code{LMI} \code{LUP} \code{LYN} \code{LYR} \code{MEN} \code{MIC} \code{MON} \code{MUS} \code{NOR} \code{OCT} \code{OPH} \code{ORI} \code{PAV} \code{PEG} \code{PER} \code{PHE} \code{PIC} \code{PSA} \code{PSC} \code{PUP} \code{PYX} \code{RET} \code{SCL} \code{SCO} \code{SCT} \code{SER} \code{SEX} \code{SGE} \code{SGR} \code{TAU} \code{TEL} \code{TRA} \code{TRI} \code{TUC} \code{UMA} \code{UMI} \code{VEL} \code{VIR} \code{VOL} \code{VUL}}
    \item{\code{NAME}}{a factor with levels \code{} \code{1} \code{101} \code{102} \code{109} \code{11} \code{110} \code{111} \code{13} \code{15} \code{16} \code{18} \code{2} \code{21} \code{23} \code{24} \code{28} \code{31} \code{36} \code{38} \code{4} \code{41} \code{43} \code{44} \code{46} \code{48} \code{5} \code{51} \code{6} \code{60} \code{67} \code{67_oph} \code{68} \code{7} \code{70} \code{70_oph} \code{72_oph} \code{8} \code{86} \code{88} \code{89} \code{98} \code{99} \code{alp_and} \code{alpha} \code{beta} \code{bet_tau} \code{chi} \code{chi^1} \code{chi^2} \code{delta} \code{epsilon} \code{eps_oph} \code{eta} \code{gamma} \code{iota} \code{kappa} \code{L} \code{L2} \code{lambda} \code{mu} \code{nu} \code{nu^3} \code{nu_oph} \code{omega} \code{omicron} \code{omicron^2} \code{p} \code{phi} \code{pi} \code{pi^2} \code{pi^3} \code{pi^4} \code{pi^5} \code{pi^6} \code{psi} \code{psi^1} \code{psi^2} \code{q} \code{rho} \code{SAO13788} \code{SAO6022} \code{sigma} \code{tau} \code{tau^1} \code{tau^2} \code{tau^3} \code{tau^4} \code{tau^5} \code{tau^6} \code{tau^8} \code{tau^9} \code{theta} \code{upsilon} \code{upsilon^1} \code{upsilon^2} \code{ups_oph} \code{xi} \code{zeta}}
    \item{\code{RA}}{a numeric vector}
    \item{\code{DEC}}{a numeric vector}
    \item{\code{MAG}}{a numeric vector}
  }
}
\details{

}
\source{

}
\references{
to be added. 
}
\examples{
data(constellation)
head(constellation)
}
\keyword{datasets}
