\name{plot_location_atlas}
\alias{plot_location_atlas}

\title{
Figure of Location for major planets 
}
\description{
Figure of Location for major planets 
}
\usage{
plot_location_atlas(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{

}
\value{
Figure
}
\references{

}
\author{
Jinlong Zhang
}
\note{

}

\seealso{

}
\examples{
JD = CAADate_DateToJD(Year = 2016, Month =1, Day =1, bGregorianCalendar=TRUE)
plot_location_atlas(JD)
}

\keyword{ apparent motion }

