\name{resstar}
\alias{resstar}
\docType{data}
\title{
Yale Bright Star Catalogue
}
\description{
Yale Bright Star Catalogue
}
\usage{data("resstar")}
\format{
  A data frame with 9110 observations on the following 4 variables.
  \describe{
    \item{\code{resRA}}{a numeric vector}
    \item{\code{rrr}}{a numeric vector}
    \item{\code{Vmag}}{a numeric vector}
    \item{\code{SAO}}{a numeric vector}
  }
}
\details{

}
\source{
   \url{http://observe.phy.sfasu.edu/SFAStarCharts/Yale/Yale.zip}
}
\references{

}
\examples{
data(resstar)
head(resstar)
}
\keyword{datasets}
