\name{ephem-package}
\alias{ephem-package}
\alias{ephem}
\docType{package}
\title{
\packageTitle{ephem}
}
\description{
\packageDescription{ephem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ephem}
\packageIndices{ephem}
creating figures for astronomical almanac
}
\author{
\packageAuthor{ephem}

Maintainer: \packageMaintainer{ephem}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
